+++ src/app/(data)/items/modItems.ts
@@ -6,6 +6,7 @@ export const modItems: ModItem[] = [
     name: 'Abrasive Rounds',
     imagePath: '/mod/abrasive_rounds.png',
     id: 'h9k5pd',
+    dlc: 'dlc2',
     description: `Imbues ammunition with BLEED and increases Lifesteal by 5% for 20s. Shots also apply BLEEDING, dealing 200 damage over 10s.`,
     wikiLinks: [`https://remnant.wiki/Abrasive_Rounds`],
   },
@@ -280,9 +281,10 @@ export const modItems: ModItem[] = [
   {
     category: 'mod',
     name: 'Cyclone',
-    imagePath: '/mod/Cyclone.png',
+    imagePath: '/mod/cyclone.png',
     id: '1bfadq',
-  description: `Maintaining an overhead Charged Melee forms a 5m sand cyclone, dealing 55 - 168 damage per second and pulls enemies towards the wielder. Each Cyclone hit applies EXPOSED for 3s. Max 15s.`,
+    dlc: 'dlc2',
+    description: `Maintaining an overhead Charged Melee forms a 5m sand cyclone, dealing 55 - 168 damage per second and pulls enemies towards the wielder. Each Cyclone hit applies EXPOSED for 3s. Max 15s.`,
     wikiLinks: [`https://remnant.wiki/Cyclone`],
     linkedItems: {
       weapon: {
@@ -308,6 +310,7 @@ export const modItems: ModItem[] = [
     name: 'Deadwood',
     imagePath: '/mod/deadwood.png',
     id: 'k7z7rz',
+    dlc: 'dlc2',
     description: `Disables automatic BLOWBACK, while increasing Thorn's Reload Speed by 25%, and Fire Rate by 15%.
 
     Reactivating the Mod will trigger an EMPOWERED BLOWBACK, which detonates all needles at once for 100% additional damage. Lasts 20s.`,